geo: split out geo helpers and move back to PX4/Firmware
 - keep map projection helpers in ecl
