+++ EKF/EKFGSF_yaw.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <geo/geo.h>
 #include <matrix/math.hpp>
 #include <mathlib/mathlib.h>
