+++ map_projection/map_projection.h
@@ -32,7 +32,7 @@
  ****************************************************************************/
 
 /**
- * @file geo.h
+ * @file map_projection.h
  *
  * Definition of geo / math functions to perform geodesic calculations
  *
@@ -47,29 +47,6 @@
 
 #include <stdbool.h>
 #include <stdint.h>
-
-static constexpr float CONSTANTS_ONE_G = 9.80665f;						// m/s^2
-
-static constexpr float CONSTANTS_STD_PRESSURE_PA = 101325.0f;					// pascals (Pa)
-static constexpr float CONSTANTS_STD_PRESSURE_KPA = CONSTANTS_STD_PRESSURE_PA / 1000.0f;	// kilopascals (kPa)
-static constexpr float CONSTANTS_STD_PRESSURE_MBAR = CONSTANTS_STD_PRESSURE_PA / 100.0f;	// Millibar (mbar) (1 mbar = 100 Pa)
-
-static constexpr float CONSTANTS_AIR_DENSITY_SEA_LEVEL_15C = 1.225f;				// kg/m^3
-static constexpr float CONSTANTS_AIR_GAS_CONST = 287.1f;					// J/(kg * K)
-static constexpr float CONSTANTS_ABSOLUTE_NULL_CELSIUS = -273.15f;				// °C
-
-static constexpr double CONSTANTS_RADIUS_OF_EARTH = 6371000;					// meters (m)
-static constexpr float  CONSTANTS_RADIUS_OF_EARTH_F = CONSTANTS_RADIUS_OF_EARTH;		// meters (m)
-
-static constexpr float CONSTANTS_EARTH_SPIN_RATE = 7.2921150e-5f;				// radians/second (rad/s)
-
-
-// XXX remove
-struct crosstrack_error_s {
-	bool past_end;		// Flag indicating we are past the end of the line/arc segment
-	float distance;		// Distance in meters to closest point on line/arc
-	float bearing;		// Bearing in radians to closest point on line/arc
-} ;
 
 /* lat/lon are in radians */
 struct map_projection_reference_s {
@@ -228,78 +205,3 @@ int globallocalconverter_toglobal(float x, float y, float z,  double *lat, doubl
  */
 int globallocalconverter_getref(double *lat_0, double *lon_0, float *alt_0);
 
-/**
- * Returns the distance to the next waypoint in meters.
- *
- * @param lat_now current position in degrees (47.1234567°, not 471234567°)
- * @param lon_now current position in degrees (8.1234567°, not 81234567°)
- * @param lat_next next waypoint position in degrees (47.1234567°, not 471234567°)
- * @param lon_next next waypoint position in degrees (8.1234567°, not 81234567°)
- */
-float get_distance_to_next_waypoint(double lat_now, double lon_now, double lat_next, double lon_next);
-
-/**
- * Creates a new waypoint C on the line of two given waypoints (A, B) at certain distance
- * from waypoint A
- *
- * @param lat_A waypoint A latitude in degrees (47.1234567°, not 471234567°)
- * @param lon_A waypoint A longitude in degrees (8.1234567°, not 81234567°)
- * @param lat_B waypoint B latitude in degrees (47.1234567°, not 471234567°)
- * @param lon_B waypoint B longitude in degrees (8.1234567°, not 81234567°)
- * @param dist distance of target waypoint from waypoint A in meters (can be negative)
- * @param lat_target latitude of target waypoint C in degrees (47.1234567°, not 471234567°)
- * @param lon_target longitude of target waypoint C in degrees (47.1234567°, not 471234567°)
- */
-void create_waypoint_from_line_and_dist(double lat_A, double lon_A, double lat_B, double lon_B, float dist,
-		double *lat_target, double *lon_target);
-
-/**
- * Creates a waypoint from given waypoint, distance and bearing
- * see http://www.movable-type.co.uk/scripts/latlong.html
- *
- * @param lat_start latitude of starting waypoint in degrees (47.1234567°, not 471234567°)
- * @param lon_start longitude of starting waypoint in degrees (8.1234567°, not 81234567°)
- * @param bearing in rad
- * @param distance in meters
- * @param lat_target latitude of target waypoint in degrees (47.1234567°, not 471234567°)
- * @param lon_target longitude of target waypoint in degrees (47.1234567°, not 471234567°)
- */
-void waypoint_from_heading_and_distance(double lat_start, double lon_start, float bearing, float dist,
-		double *lat_target, double *lon_target);
-
-/**
- * Returns the bearing to the next waypoint in radians.
- *
- * @param lat_now current position in degrees (47.1234567°, not 471234567°)
- * @param lon_now current position in degrees (8.1234567°, not 81234567°)
- * @param lat_next next waypoint position in degrees (47.1234567°, not 471234567°)
- * @param lon_next next waypoint position in degrees (8.1234567°, not 81234567°)
- */
-float get_bearing_to_next_waypoint(double lat_now, double lon_now, double lat_next, double lon_next);
-
-void get_vector_to_next_waypoint(double lat_now, double lon_now, double lat_next, double lon_next, float *v_n, float *v_e);
-
-void get_vector_to_next_waypoint_fast(double lat_now, double lon_now, double lat_next, double lon_next, float *v_n, float *v_e);
-
-void add_vector_to_global_position(double lat_now, double lon_now, float v_n, float v_e, double *lat_res, double *lon_res);
-
-int get_distance_to_line(struct crosstrack_error_s *crosstrack_error, double lat_now, double lon_now,
-				  double lat_start, double lon_start, double lat_end, double lon_end);
-
-int get_distance_to_arc(struct crosstrack_error_s *crosstrack_error, double lat_now, double lon_now,
-				 double lat_center, double lon_center,
-				 float radius, float arc_start_bearing, float arc_sweep);
-
-/*
- * Calculate distance in global frame
- */
-float get_distance_to_point_global_wgs84(double lat_now, double lon_now, float alt_now,
-		double lat_next, double lon_next, float alt_next,
-		float *dist_xy, float *dist_z);
-
-/*
- * Calculate distance in local frame (NED)
- */
-float mavlink_wpm_distance_to_point_local(float x_now, float y_now, float z_now,
-		float x_next, float y_next, float z_next,
-		float *dist_xy, float *dist_z);
